f: Implement Unity 6 AndroidManifest initialization from UnityManifest template
The UnityManifest in Unity 6 contains 2 activities. The normal one as
Unity did before and the new GameActivity.
But building of APKs requires that only one is present.
To make this easy on the game developer, we check the Application Entry
point and remove the other activity. If both entry poitns are active we
remove the GameActivity.