f: Implement platform module validation for web and android
Unity does not provide API to check if a platform is present before
switching platforms. It only fails if not present.

To overcome this we use reflection to call Unity's internal API from the
ModuleManager to query the module presence state
(IsPlatformSupportLoaded - we do it ByBuildTarget for typing reasons)
before doing the platform switch or to inform the developer that no
supported module is found and that AirConsole can not be used.